+++ spec/rubocop/formatter/disabled_config_formatter_spec.rb
@@ -242,4 +242,46 @@ RSpec.describe RuboCop::Formatter::DisabledConfigFormatter, :isolated_environmen
       expect(output.string).to eq(heading)
     end
   end
+
+  context 'with auto-correct supported cop' do
+    before do
+      module Test
+        class Cop3 < ::RuboCop::Cop::Cop
+          def autocorrect
+            # Dummy method to respond to #support_autocorrect?
+          end
+        end
+      end
+
+      formatter.started(['test_auto_correct.rb'])
+      formatter.file_started('test_auto_correct.rb', {})
+      formatter.file_finished('test_auto_correct.rb', offenses)
+      formatter.finished(['test_auto_correct.rb'])
+    end
+
+    let(:expected_rubocop_todo) do
+      [heading,
+       '# Offense count: 1',
+       '# Cop supports --auto-correct.',
+       'Test/Cop3:',
+       '  Exclude:',
+       "    - 'test_auto_correct.rb'",
+       ''].join("\n")
+    end
+
+    let(:offenses) do
+      [
+        RuboCop::Cop::Offense.new(
+          :convention,
+          location,
+          'message',
+          'Test/Cop3'
+        )
+      ]
+    end
+
+    it 'adds a comment about --auto-correct option' do
+      expect(output.string).to eq(expected_rubocop_todo)
+    end
+  end
 end